ITG3200 and ADXL345 working. Printing data for test

+++ src/Arduino - AS6/ScoeComms/Utils/I2CHelpers.h
@@ -9,7 +9,7 @@
 #ifndef I2CHELPERS_H_
 #define I2CHELPERS_H_
 
-void readI2CBytes(uint8_t device, uint8_t address, uint8_t *buffer, int offset, int count);
+void readI2CBytes(uint8_t device, uint8_t address, uint8_t *buffer, int offset, uint8_t count);
 uint8_t writeI2CByte(uint8_t device, uint8_t address, uint8_t val);
 
 #endif /* I2CHELPERS_H_ */
